Added tests for timingSafeCompare, moved implementation to DataExtensions
